renamed files to conform standard; fixed issues remaining from previous refactorings
