+++ sote/libsote/rock-layers.lua
